Refactor statistics collection logic into a separate module
